perf: compact indexed timestamps to 4-byte Unix seconds
Size/date sorting and filtering need fixed-width columns for fast random
access, ruling out variable-length encoding, but the timestamps' 100ns
FILETIME precision is unused by a search tool. Storing whole-second Unix
time (FileTimeHelper) instead cuts each timestamp column from 8 to 4 bytes,
dropping the memory/disk growth from the previous change from roughly +97%
to +61% per row with no loss of sort/filter capability.

+++ Core/SearchIndex/RecordIndex/RuntimeIndexCacheExtensions.cs
@@ -111,9 +111,9 @@ internal static class RuntimeIndexCacheExtensions
                 var parentIdHigh = recordReader.ReadUInt64();
                 var flags = (FileRecordFlags)recordReader.ReadUInt16();
                 var size = recordReader.ReadInt64();
-                var creationTimeUtc = recordReader.ReadInt64();
-                var lastWriteTimeUtc = recordReader.ReadInt64();
-                var lastAccessTimeUtc = recordReader.ReadInt64();
+                var creationTimeUtc = recordReader.ReadUInt32();
+                var lastWriteTimeUtc = recordReader.ReadUInt32();
+                var lastAccessTimeUtc = recordReader.ReadUInt32();
 
                 var id = new UInt128(idHigh, idLow);
                 var parentId = new UInt128(parentIdHigh, parentIdLow);
